requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
dlx==1.0.4
docplex==2.15.194
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.14.5
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
more-itertools==8.5.0
mpmath==1.1.0
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
ntlm-auth==1.5.0
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psutil==5.7.2
pybind11==2.6.0
pycparser==2.20
pydot==1.4.1
Pygments==2.7.1
pylatexenc==2.7
pyparsing==2.4.7
pyrsistent==0.17.3
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qiskit==0.23.0
qiskit-aer==0.7.0
qiskit-aqua==0.8.0
qiskit-ibmq-provider==0.11.0
qiskit-ignis==0.5.0
qiskit-terra==0.16.0
Quandl==3.5.3
requests==2.24.0
requests-ntlm==1.1.0
retworkx==0.5.0
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sympy==1.6.2
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
yfinance==0.1.55
zipp==3.3.1